Slice `expression` mod into `pattern/expressions` and `builder/expressions`
