+++ rust/pattern/variable/builder/expression.rs
@@ -20,23 +20,11 @@
  *
  */
 
-mod constant;
-mod function;
-mod operation;
-mod parenthesis;
-
-use std::{fmt, iter};
-
-use chrono::NaiveDateTime;
-pub use constant::Constant;
-pub use function::Function;
-pub use operation::Operation;
-pub use parenthesis::Parenthesis;
-
 use crate::{
     common::token::{Function as FunctionToken, Operation as OperationToken},
     pattern::{Reference, UnboundConceptVariable, UnboundValueVariable, UnboundVariable},
 };
+use crate::pattern::expression::{Expression, Function, Operation};
 
 pub trait ExpressionBuilder {
     fn add(self, right: impl Into<Expression>) -> Expression;
@@ -106,87 +94,6 @@ impl<T: Into<Expression>> ExpressionBuilder for T {
 
     fn round(arg: impl Into<Expression>) -> Expression {
         Expression::Function(Function { function_name: FunctionToken::Round, args: vec![Box::from(arg.into())] })
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub enum Expression {
-    Operation(Operation),
-    Function(Function),
-    Constant(Constant),
-    Parenthesis(Parenthesis),
-    Variable(UnboundVariable),
-}
-
-impl fmt::Display for Expression {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Expression::Operation(operation) => write!(f, "{operation}"),
-            Expression::Function(function) => write!(f, "{function}"),
-            Expression::Constant(constant) => write!(f, "{constant}"),
-            Expression::Parenthesis(parenthesis) => write!(f, "{parenthesis}"),
-            Expression::Variable(variable) => write!(f, "{variable}"),
-        }
-    }
-}
-
-impl Expression {
-    pub fn references_recursive(&self) -> Box<dyn Iterator<Item = &Reference> + '_> {
-        match self {
-            Expression::Operation(operation) => operation.references_recursive(),
-            Expression::Function(function) => function.references_recursive(),
-            Expression::Constant(_constant) => Box::new(iter::empty()),
-            Expression::Parenthesis(parenthesis) => parenthesis.references_recursive(),
-            Expression::Variable(variable) => variable.references(),
-        }
-    }
-}
-
-impl From<UnboundValueVariable> for Expression {
-    fn from(variable: UnboundValueVariable) -> Self {
-        Self::Variable(variable.into())
-    }
-}
-
-impl From<UnboundConceptVariable> for Expression {
-    fn from(variable: UnboundConceptVariable) -> Self {
-        Self::Variable(variable.into())
-    }
-}
-
-impl From<i64> for Expression {
-    fn from(value: i64) -> Self {
-        Self::Constant(value.into())
-    }
-}
-
-impl From<f64> for Expression {
-    fn from(value: f64) -> Self {
-        Self::Constant(value.into())
-    }
-}
-
-impl From<bool> for Expression {
-    fn from(value: bool) -> Self {
-        Self::Constant(value.into())
-    }
-}
-
-impl From<String> for Expression {
-    fn from(value: String) -> Self {
-        Self::Constant(value.into())
-    }
-}
-
-impl From<&str> for Expression {
-    fn from(value: &str) -> Self {
-        Self::Constant(value.into())
-    }
-}
-
-impl From<NaiveDateTime> for Expression {
-    fn from(value: NaiveDateTime) -> Self {
-        Self::Constant(value.into())
     }
 }
